+++ Kernel/include/Tier0/kstdio.h
@@ -18,7 +18,7 @@ void kprint(const s8 *szString);
 void kputi(s32 Number);
 void kprintf(const s8 *Format, ...);
 void kdump(u8 *bData, u32 Length);
-void kprint_hex(u32 Number);
+void kprint_hex(u64 Number);
 void kstdio_set_globals(u8 line, u8 cur_x, u8 cur_y);
 s32 kmemcmp(u8 *MemA, u8 *MemB, u32 Length);
